[PF-2364] Bundle the command docs in release
Add picocli command doc generation to gradle and
include the doc artifacts in the release distribution
bundle. Also update docs for a few commands.

This change is based on the gj/doc-gen branch. But at this
point it only has the command doc generation stuff. The readme
doc and some additional instructions in the original branch
need further cleanup and breaking into separate articles.