Update documentation to correctly describe default instance ID (#500)
* Generate instance ID if one is not provided

If no instance ID is provided when the command '\''terra resource create gcp-notebook'\'' is run, generate one from user email and time.

* Refactor test to check for correct instance ID

Refactor test to validate that generated instance ID has user email as prefix. Revert changes to command as changes will be made in WSM instead.

* Modify WSM service to generate instance ID

When no instnace ID is provided, the WSM generates an instance ID from
the user email and the date/time.

* Revert changes to command to create GCP notebook

Revert changes to files as, given the nature of the restrictions of the instance ID (only lowercase alphanumeric and hyphens allowed), leveraging the user's email for the instance ID doesn't make sense. Therefore, we'll change the documentation to reflect the actual implementation.

* Refactor documentation to reflect default instance ID

The default instance ID is the provided name of the resource, with
underscores removed and all letters lowercase.

* Reorder publish workflow (#502)

* bump 0.356.0

* Reword the content to be future-proof

Reword content to apply in future situations if the instance ID
generation changes.

---------

Co-authored-by: Dexter Amundsen <dexamundsen@verily.com>
Co-authored-by: bumptagbot <robot@terra.team>